Break out loader from project
