+++ internal/cmd/run.go
@@ -75,7 +75,8 @@ func runCmd() *cobra.Command {
 
 			{
 			searchBlocks:
-				blocks, err := loadTasks(proj, cmd.OutOrStdout(), cmd.InOrStdin(), true)
+				loader := getLoader(cmd)
+				blocks, err := loader.LoadTasks(proj, fAllowUnknown, fAllowUnnamed, true)
 				if err != nil {
 					return err
 				}
